Add enum refs to sys::Enum types
I think ideally the syntax should allow something like @WeatherCondEnum, but then we have
what looks like a data instance ref in a Spec slot as the scalar default.  Its kind of
a weird case to handle in the compiler.  So for now just use the Ref "xxx" syntax (but we
don't have any resolution checking in the compiler)